+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Not released yet
 ## Popup
 - Added a delete option to the blacklisted sites list within settings for every blacklisted site
 - The button that feeds the list of blacklisted sites now has the number of blacklisted sites
-- Small changes in the style of settings options
+- Changes in the style of the options in settings
 
 ---